android打包配置完成，成功编译出APK

+++ PopStar3.5/Classes/connection/WebSocketConnection.cpp
@@ -87,7 +87,7 @@ void WebSocketConnection::onClose(cocos2d::network::WebSocket* ws)
 
 void WebSocketConnection::onError(cocos2d::network::WebSocket* ws, const cocos2d::network::WebSocket::ErrorCode& error)
 {
-	log("Error was fired, error code: %d", error);
+	log("Error was fired, error code: %d", (int)error);
 }
 
 void WebSocketConnection::disconnect()
